Creates SiftImpl class, moves API to Sift class
