+++ hello-sift/build.gradle
@@ -38,6 +38,7 @@ dependencies {
     compile 'com.android.support:appcompat-v7:25.0.0'
     compile project(':sift')
 
+    compile 'com.android.support.constraint:constraint-layout:1.0.2'
     androidTestCompile 'com.android.support:support-annotations:25.0.0'
     androidTestCompile 'com.android.support.test:runner:0.5'
     androidTestCompile 'com.android.support.test:rules:0.5'
