+++ hello-sift/src/main/AndroidManifest.xml
@@ -11,9 +11,11 @@
         <activity android:name=".HelloSift">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".OtherActivity"></activity>
     </application>
 
 </manifest>
