+++ sift/src/main/java/siftscience/android/AppStateCollector.java
@@ -42,7 +42,7 @@ public class AppStateCollector implements LocationListener,
         GoogleApiClient.ConnectionCallbacks,
         GoogleApiClient.OnConnectionFailedListener {
     private static final String TAG = AppStateCollector.class.getSimpleName();
-    private final Sift sift;
+    private final SiftImpl sift;
     private final Context context;
 
     private String activityClassName;
@@ -53,7 +53,7 @@ public class AppStateCollector implements LocationListener,
     private Location location;
     private Location lastLocation;
 
-    public AppStateCollector(Sift sift, Context context) {
+    public AppStateCollector(SiftImpl sift, Context context) {
         this.sift = sift;
         this.context = context.getApplicationContext();
